Update Dockerfile w.r.t. flask 2.2.0
In the latest version - 2.2.0 - the flask developers decided that it is over-complicated to have different environment variables for more or less the same feature - development and debugging. Therefore, the default server - Werkzeug - shall be used until the development process is complete. Once the app is production-ready then we can easily switch the server to waitress.